# api.yaml
openapi: 3.0.0

info:
  title: My API
  version: 1.0.0
  description: User API with complex data types (MongoDB _id as string in JSON)

servers:
  - url: /
    description: Same host as Swagger UI (local or Render)

components:
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          description: MongoDB ObjectId (24 hex characters)
          example: 507f1f77bcf86cd799439011

        name:
          type: string
          example: Manoj

        age:
          type: integer
          example: 25

        isActive:
          type: boolean
          example: true

        skills:
          type: array
          items:
            type: string
          example: ["Node.js", "React", "MongoDB"]

        address:
          type: object
          properties:
            city:
              type: string
              example: Bangalore
            pincode:
              type: integer
              example: 560001

        image:
          type: string
          format: uri
          description: Profile photo URL (served from Unsplash CDN)
          example: https://images.unsplash.com/photo-1494790108377-be9c29b29330?auto=format&fit=crop&w=400&q=80

        bio:
          type: string
          description: Short profile summary
          example: Full-stack engineer building APIs and UIs.

        imageCredit:
          type: string
          description: Photographer attribution when the image comes from the Unsplash API
          example: Photo by Alex Doe on Unsplash

    Error:
      type: object
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          example: Error message

paths:
  /users:
    get:
      summary: Get all users
      description: Fetch all users with detailed data (each includes _id)

      responses:
        200:
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
              example:
                - _id: 507f1f77bcf86cd799439011
                  name: Manoj
                  age: 25
                  isActive: true
                  skills: ["Node.js", "React"]
                  address:
                    city: Bangalore
                    pincode: 560001
                  image: https://images.unsplash.com/photo-1494790108377-be9c29b29330?auto=format&fit=crop&w=400&q=80
                  bio: API designer and Node enthusiast.

                - _id: 507f191e810c19729de860ea
                  name: Kumar
                  age: 28
                  isActive: false
                  skills: ["Java", "Spring"]
                  address:
                    city: Mysore
                    pincode: 570001
                  image: https://images.unsplash.com/photo-1507003211169-0a1dd7228f2d?auto=format&fit=crop&w=400&q=80
                  bio: Backend engineer focused on Spring services.

        404:
          description: No users found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Create user
      description: Add a new user (MongoDB assigns _id)

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
            example:
              name: Rahul
              age: 22
              isActive: true
              skills: ["Python"]
              address:
                city: Delhi
                pincode: 110001
              image: https://images.unsplash.com/photo-1500648767791-00dcc994a43e?auto=format&fit=crop&w=400&q=80
              bio: Data and Python tooling.

      responses:
        201:
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: MongoDB user _id (24 hex ObjectId string)
        schema:
          type: string
          pattern: '^[a-fA-F0-9]{24}$'
          example: 507f1f77bcf86cd799439011

    put:
      summary: Update user
      description: Update a user by _id. Do not rely on _id in body; path id wins. _id in body is ignored.

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
            example:
              name: Rahul Kumar
              age: 23
              isActive: true
              skills: ["Python", "Django"]
              address:
                city: Delhi
                pincode: 110001
              image: https://images.unsplash.com/photo-1500648767791-00dcc994a43e?auto=format&fit=crop&w=400&q=80
              bio: Updated profile.

      responses:
        200:
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

        400:
          description: Invalid id or missing name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete user
      description: Remove a user by _id

      responses:
        204:
          description: User deleted (no response body)

        400:
          description: Invalid user id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"